test: cover ambiguous compatibility recovery and custom index discovery
